domain/src/main/java/com/gsv/codeflix/admin/catalog/domain/category/Category.java
package com.gsv.codeflix.admin.catalog.domain.category;

import java.time.Instant;
import java.util.UUID;

public class Category {
    public String id;
    private final String name;
    private final String description;
    private final boolean isActive;
    private final Instant createdAt;
    private final Instant updatedAt;
    private final Instant deletedAt;

    private Category(String id, String name, String description, boolean isActive, Instant createdAt, Instant updatedAt, Instant deletedAt) {
        this.id = id;
        this.name = name;
        this.description = description;
        this.isActive = isActive;
        this.createdAt = createdAt;
        this.updatedAt = updatedAt;
        this.deletedAt = deletedAt;
    }

    public static Category newCategory(String name, String desc, boolean isActive) {
        final var id = UUID.randomUUID().toString();
        final var timestamp = Instant.now();
        return new Category(id, name, desc, isActive, timestamp, timestamp, null);
    }

    public String getId() {
        return id;
    }

    public String getName() {
        return name;
    }

    public String getDescription() {
        return description;
    }

    public boolean getIsActive() {
        return isActive;
    }

    public Instant getCreatedAt() {
        return createdAt;
    }

    public Instant getUpdatedAt() {
        return updatedAt;
    }

    public Instant getDeletedAt() {
        return deletedAt;
    }
}